get_journal/requirements.txt
functions-framework==3.*
google-cloud-bigquery
google-generativeai
requests
